backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
google-cloud-aiplatform
google-cloud-speech
sympy
httpx
